[move][cli] Support safe module republishing with breaking changes checks
- If the CLI is trying to compile `M.move`, the logic for finding compilation deps now removes `M.move` from the deps list.
More generally, any files directly or transitively reachable from the compilation targets are removed from the deps list. This
fixes the annoying "duplicate definition" warning that you used to get when trying to re-publish a module
- Before a `move publish`, the CLI runs the bytecode linking and data layout compatibility checks and reports an error if the
code to be published could cause a breaking change to either linking or data layout
- Added tests for republishing, running `move check M.move` after publishing `M.move`, and the breaking changes checks

Closes: #6753
