+++ language/tools/move-cli/tests/testsuite/data_breaking_change/M.move
@@ -1,0 +1,5 @@
+address 0x2 {
+module M {
+    resource struct R { f: u64, g: u64 } // extra field g
+}
+}
