+++ language/tools/move-cli/tests/testsuite/data_breaking_change/args.exp
@@ -1,0 +1,5 @@
+Command `publish src/modules/`:
+Command `publish M.move`:
+Breaking change detected--publishing aborted. Re-run with --ignore-breaking-changes to publish anyway.
+Error: Layout API for structs of module 00000000000000000000000000000002::M has changed. Need to do a data migration of published structs
+Command `publish --ignore-breaking-changes M.move`:
