+++ language/tools/move-cli/tests/testsuite/data_breaking_change/args.txt
@@ -1,0 +1,3 @@
+publish src/modules/
+publish M.move
+publish --ignore-breaking-changes M.move
