+++ language/tools/move-cli/tests/testsuite/linking_breaking_change/src/modules/M.move
@@ -1,0 +1,6 @@
+address 0x2 {
+module M {
+    public fun f() {}
+    public fun g() {}
+}
+}
