+++ language/tools/move-cli/tests/testsuite/publish_then_check/args.exp
@@ -1,0 +1,2 @@
+Command `publish src/modules`:
+Command `check`:
