+++ language/tools/move-cli/tests/testsuite/publish_then_check/args.txt
@@ -1,0 +1,2 @@
+publish src/modules
+check
