+++ language/tools/move-cli/tests/testsuite/publish_then_check/src/modules/M.move
@@ -1,0 +1,5 @@
+address 0x2 {
+module M {
+    public fun f() {}
+}
+}
