+++ language/tools/move-cli/tests/testsuite/publish_then_run/args.exp
@@ -1,0 +1,2 @@
+Command `publish src/modules`:
+Command `run src/scripts/main.move`:
