+++ language/tools/move-cli/tests/testsuite/publish_then_run/args.txt
@@ -1,0 +1,2 @@
+publish src/modules
+run src/scripts/main.move
