+++ language/tools/move-cli/tests/testsuite/publish_then_run/src/scripts/main.move
@@ -1,0 +1,3 @@
+script {
+fun main() {}
+}
