+++ language/tools/move-cli/tests/testsuite/republish_module/args.exp
@@ -1,0 +1,2 @@
+Command `publish src/modules`:
+Command `publish src/modules`:
