+++ language/tools/move-cli/tests/testsuite/republish_module/args.txt
@@ -1,0 +1,2 @@
+publish src/modules
+publish src/modules
