+++ language/tools/move-cli/tests/testsuite/republish_module/src/modules/M.move
@@ -1,0 +1,5 @@
+address 0x2 {
+module M {
+    public fun f() {}
+}
+}
