+++ language/vm/src/lib.rs
@@ -10,6 +10,7 @@ use std::fmt;
 
 pub mod access;
 pub mod check_bounds;
+pub mod compatibility;
 #[macro_use]
 pub mod errors;
 pub mod constant;
